requirements.txt
pytest==7.3.1
requests==2.25.1
PyVirtualDisplay==2.0
allure-pytest==2.8.32
allure-python-commons==2.8.32
prometheus-client==0.9.0
selenium==4.28.1